\name{DATA.DIRECTORY}
\alias{DATA.DIRECTORY}
\docType{data}
\title{Directory for the final data storage
 
}
\description{Finalized inventories and data files are written here
 
}
\usage{ DATA.DIRECTORY }
\format{
  The format is:
 chr "DataDirectory"
}
\details{When data is processed from the web or local csv files
  files are created and put in this directory
 
}
 
\keyword{datasets}
