\name{FORMAT.URL}
\alias{FORMAT.URL}
\docType{data}
\title{ A string for completing the http request
 
}
\description{htpp requests for data are made by combining 5 strings:
  a base string, a year string, a station webid, a year, and a format
  string
 
}
\usage{ FORMAT.URL }
\format{
  The format is:
 chr "&Month=1&Day=1&format=csv"
}
\details{ A string to specify that we want monthly data in a csv file
 
}
\source{\url{http://climate.weatheroffice.gc.ca}
 
}
 
\keyword{datasets}
