\name{MASTER.STATION.LIST}
\alias{MASTER.STATION.LIST}
\docType{data}
\title{Name for the local version of the master station list
 
}
\description{defines the name of the local version of the master list
 
}
\usage{ MASTER.STATION.LIST }
\format{
  The format is:
 chr "EnvCanadaMaster.csv"
}
\details{ Name of the local file. used in other functions
 
}
 
\keyword{datasets}
