\name{MONTHLY.STATION.LIST}
\alias{MONTHLY.STATION.LIST}
\docType{data}
\title{Name of the local file for Monthly stations
 
}
\description{This file will contain a subset of master list.
  feilds have been edited to make scraping easier. 
  Only stations with monthly data are in the list
 
}
\usage{ MONTHLY.STATION.LIST }
\format{
  The format is:
 chr "MonthlyStations.Env.csv"
}
\details{ The name used for the local version of the stations
  that report monthly
 
}
 
\keyword{datasets}
