\name{STARTING.STATION.ID}
\alias{STARTING.STATION.ID}
\docType{data}
\title{The starting unique Id created for these stations
 
}
\description{Stations are numbered sequentially when they are taken
  out of the master list. Obviously if the master list changes this
  unquie Id will change. It is used only for keeping track
  of files during processing. It is not an officially recognized
  Id and should not be confused with WMo numbers or the like
 
}
\usage{ STARTING.STATION.ID }
\format{
  The format is:
 num 99111111
}
\details{ Stations are numbered sequentially 
 
}
 
 
\examples{
data(STARTING.STATION.ID)
print(STARTING.STATION.ID)
}
\keyword{datasets}
