\name{STATION.URL}
\alias{STATION.URL}
\docType{data}
\title{ url for getting master station list
 
}
\description{ url to get csv file of master list
 
}
\usage{ STATION.URL }
\format{
  The format is:
 chr "http://scraperwiki.com/api/1.0/datastore/
      sqlite?format=csv&name=can-weather-stations&query=select+*+from+`swdata`"
}
\details{ Output of the scraper written by DrJ of clearclimatecode.org
 
}
\source{\url{http://scraperwiki.com/scrapers/can-weather-stations/}
 
}
 
 
\keyword{datasets}
