\name{YEAR.URL}
\alias{YEAR.URL}
\docType{data}
\title{ A string for making the http request
 
}
\description{ 1 of 5 strings concatenated to make the final request
 
}
\usage{ YEAR.URL }
\format{
  The format is:
 chr "&Year="
}
\details{ used to make the http request  
 
}
 
 
\keyword{datasets}
